mini-workshops/cdm_detector_example/requirements.txt
jupyter
pyyaml
nbstripout
matplotlib
numpy
scipy
tqdm
flake8